Refs #117 added check if instance is master
